Re-add gestures to date label
