+++ app/src/main/res/values/strings.xml
@@ -36,7 +36,7 @@
     <string name="setting_general_title">General</string>
 
     <string name="setting_enable_gestures">Gestures</string>
-    <string name="setting_enable_gestures_summary">Allow swiping to change current day</string>
+    <string name="setting_enable_gestures_summary">Allow swiping date to change current day</string>
     <string name="setting_clip_accomplishments">Auto clipping</string>
     <string name="setting_clip_accomplishments_summary">Remove empty lines from Accomplishment posts</string>
     <string name="setting_time_selection">Time Selection</string>
